Performing Whois lookups async | dictionary to retry failed lookups
In this commit I changed the Whois lookup to be async and throw the data inside the session permutations, so whenever we send request to /domains we get the response immediately and the Whois lookups will be added once its finished (so u will see it in the next request ).

Also I noticed that some of the Whois lookups returns null so I made a dictionary to store the null record and retry to get the data by sending the request again.